create class Person, create module phase_settable, update classes to DRY
